Add modifications

+++ src/main/java/com/threedgarden/api/service/InventoryService.java
@@ -10,24 +10,45 @@ import java.util.Optional;
 
 @Service
 public class InventoryService {
-
+    private final InventoryRepository inventoryRepository;
     @Autowired
-    private InventoryRepository inventoryRepository;
+    public InventoryService(InventoryRepository inventoryRepository){
+        this.inventoryRepository = inventoryRepository;
+    }
 
     public List<Inventory> getAllInventories() {
         return inventoryRepository.findAll();
     }
 
-    public Optional<Inventory> getInventoryById(Long id) {
-        return inventoryRepository.findById(id);
+    public Inventory getInventoryById(Long id){
+        return inventoryRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("El producto con el " + id + " no existe en el inventario")
+        );
     }
 
-    public Inventory saveInventory(Inventory inventory) {
+    public Inventory addInventory(Inventory inventory){
         return inventoryRepository.save(inventory);
     }
 
-    public void deleteInventory(Long id) {
-        inventoryRepository.deleteById(id);
+    public Inventory deleteInventoryById(Long id) {
+        Inventory tmp = null;
+        if(inventoryRepository.existsById(id)){
+            tmp = inventoryRepository.findById(id).get();
+            inventoryRepository.deleteById(id);
+            return tmp;
+        }
+        return tmp;
+    }
+
+    public Inventory updateInventory(Long id, Inventory inventory) {
+        Inventory existingInventory = inventoryRepository.findById(id).orElseThrow(() -> new RuntimeException("Inventario no encontrado"));
+
+        existingInventory.setStatus(inventory.getStatus());
+        existingInventory.setQuantity(inventory.getQuantity());
+        existingInventory.setRegistrationDate(inventory.getRegistrationDate());
+
+        return inventoryRepository.save(existingInventory);
+
     }
